% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EnsembleFeaturesRanking.R
\name{ensemble.features.ranking}
\alias{ensemble.features.ranking}
\title{Ensemble Features Ranking}
\usage{
ensemble.features.ranking(dataset, ...)
}
\arguments{
\item{dataset}{dataset with numerial features and factor class}

\item{nRows}{number for each sample run, 10 default value}

\item{nTries}{number of runs, ncol(dataset) / 2, default value}

\item{cutOff}{k = n (selects n features); k=0 (selects all features); k=-1 (automatic selecion) default}

\item{methods}{list for relevancy metric calculation}
}
\value{
Ranked features list sorted by relevancy
}
\description{
Takes a dataset and outputs a features name list sorted by relevancy
}
\examples{
\dontrun{
}
#' data(allDataProbe)
set.seed(134)
methods <- list(gain.ratio,chi.squared)
features <- ensemble.features.ranking(allDataProbe, nRows=nrow(allDataProbe)/2, nTries=20, cutOff= -1, methods=methods)
print(features)

}
